MaxPool is now storing Coords
Max Pool is now storing the Coords of the selected input Pixels for use in backprop and the last written Pixels of LossInput for ease of cleanup in the next image. ALSO: FIXED A BIG BUG IN MAXPOOL, WHERE THE MAX FROM EVERY ROW WAS SELECTED AND WRITTEN INTO OUTPUT, CAUSING ONLY PIXELS IN THE LAST ROW OF THE WINDOW TO MATTER. PLEASE DONT DELETE BRACKETS "{" AND "}". IT DOES NOT MAKE THE CODE MORE READABLE OR PRETTIER AND ONLY INTRODUCES BUGS.